- Fixed: Crash when loading a rom with an unknown file extension. - Fixed: Crash when stopping a game. - Improved: DirectX Plugin.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@237 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Core/Common/Src/FileUtil.cpp
@@ -28,7 +28,7 @@ bool File::IsDirectory(const std::string &filename) {
 
 std::string SanitizePath(const std::string &filename) {
 	std::string copy = filename;
-	for (int i = 0; i < copy.size(); i++)
+	for (size_t i = 0; i < copy.size(); i++)
 		if (copy[i] == '/')
 			copy[i] = '\\';
 	return copy;
